v2.6.4: fixed crash on seemingly random Windows systems
* disabled AAX, enabled SSE2

* VST3 plug-in

* support 64-bit signal chains

* optimized dithering

* updated documentation

* re-factored and optimized code

* updated JUCE framework to 5.3.2

* updated VST library to 3.6.8

* show CPU information on start-up

* use dynamic runtime linking on Windows

* use Clang on Linux (C++14)

* support Visual Studio 2017

+++ Source/plugin_processor.cpp
@@ -718,7 +718,8 @@ void KmeterAudioProcessor::processBlock(
             {
                 outputGain_ = static_cast<double>(
                                   MeterBallistics::decibel2level(
-                                      currentAttenuationDecibel_));
+                                      static_cast<float>(
+                                          currentAttenuationDecibel_)));
             }
             else
             {
@@ -739,7 +740,8 @@ void KmeterAudioProcessor::processBlock(
             {
                 outputGain_ = static_cast<double>(
                                   MeterBallistics::decibel2level(
-                                      currentAttenuationDecibel_));
+                                      static_cast<float>(
+                                          currentAttenuationDecibel_)));
             }
             else
             {
@@ -910,7 +912,8 @@ void KmeterAudioProcessor::processBlock(
             {
                 outputGain_ = static_cast<double>(
                                   MeterBallistics::decibel2level(
-                                      currentAttenuationDecibel_));
+                                      static_cast<float>(
+                                          currentAttenuationDecibel_)));
             }
             else
             {
@@ -931,7 +934,8 @@ void KmeterAudioProcessor::processBlock(
             {
                 outputGain_ = static_cast<double>(
                                   MeterBallistics::decibel2level(
-                                      currentAttenuationDecibel_));
+                                      static_cast<float>(
+                                          currentAttenuationDecibel_)));
             }
             else
             {
